Update zarr-python represenation on ZIC
Welcome to @jhamman as a new ZIC representation for zarr-python and @d-v-b as his second in cases where he is unavailable for a vote. This change matches the development of a new ZEP1 implementation on the `v3` branch of the zarr-python repo, which re-integrates work from the zarrita implementation by @normanrz and others. 👏🏽  Very much looking forward to the upcoming release.

Many thanks to @grlee77 for his contributions to date.

Internal reference: https://groups.google.com/g/zarr-python-core-devs/c/sHsbWBOBHSs?pli=1